+++ tinyspline/src/main.cpp
@@ -1,12 +1,10 @@
 #include "fmt/format.h"
 #include "fmt/ranges.h" // support printing of (nested) containers & tuples
 
-#include <vector>
 #include "tinysplinecxx.h"
+#include <vector>
 
 #include "hd/hd_keypress.hpp"
-
-
 
 
 // using namespace fmt::literals; // just to make the format literals visible
@@ -58,32 +56,35 @@ int main()
 
     // setting up a two-dimensional cubic spline with 4 control points
     // BSline(#Ctrl-pts, dimension, degree)
-    constexpr int num_ctrl_pts = 4;  // no. of control points
-    constexpr int ndim = 2;          // dimension of control points
-    constexpr int degree = 3;        // degree curve
+    constexpr int num_ctrl_pts = 4; // no. of control points
+    constexpr int ndim = 2;         // dimension of control points
+    constexpr int degree = 3;       // degree curve
 
-    tinyspline::BSpline spline = tinyspline::BSpline(num_ctrl_pts,ndim,degree);
+    tinyspline::BSpline spline = tinyspline::BSpline(num_ctrl_pts, ndim, degree);
     std::vector<tinyspline::real> ctrl_pts = spline.controlPoints();
-    ctrl_pts[0] = 0; ctrl_pts[1] = 0;  // P0
-    ctrl_pts[2] = 2; ctrl_pts[3] = 3;  // P1
-    ctrl_pts[4] = 7; ctrl_pts[5] = 4;  // P2
-    ctrl_pts[6] = 6; ctrl_pts[7] = 1;  // P3
+    ctrl_pts[0] = 0;
+    ctrl_pts[1] = 0; // P0
+    ctrl_pts[2] = 2;
+    ctrl_pts[3] = 3; // P1
+    ctrl_pts[4] = 7;
+    ctrl_pts[5] = 4; // P2
+    ctrl_pts[6] = 6;
+    ctrl_pts[7] = 1; // P3
     spline.setControlPoints(ctrl_pts);
 
     // sample line in npts different points
     constexpr int npts = 11;
     std::vector<tinyspline::real> pts = spline.sample(npts);
 
-    for (int i=0; i<pts.size()/ndim-1; ++i)
-    {
-        tinyspline::real px0 = pts[i*ndim];
-        tinyspline::real py0 = pts[i*ndim + 1];
-        tinyspline::real px1 = pts[(i+1)*ndim];
-        tinyspline::real py1 = pts[(i+1)*ndim + 1];
+    for (size_t i = 0; i < pts.size() / ndim - 1; ++i) {
+        tinyspline::real px0 = pts[i * ndim];
+        tinyspline::real py0 = pts[i * ndim + 1];
+        tinyspline::real px1 = pts[(i + 1) * ndim];
+        tinyspline::real py1 = pts[(i + 1) * ndim + 1];
 
         // draw line from p0 to p1 => line(p0,p1)
-        fmt::print("i={:2}: p0x={:.2f}, p0y={:.2f}, p1x={:.2f}, p1y={:.2f}\n",
-        i, px0, py0, px1, py1);
+        fmt::print("i={:2}: p0x={:.2f}, p0y={:.2f}, p1x={:.2f}, p1y={:.2f}\n", i, px0,
+                   py0, px1, py1);
     }
 
     //   fmt::print(" x = {:.2f}\n", fmt::join(x, ", "));
